Deleting the TaskConfig, so it can be reborn
